Change base conversion implementation
Changed implementation from using the direct conversion algorithm to for
numbers in base 2, 8 and 16 to doing it in two steps: first convert to
base 10, then to desired destination base for the number.

Also added input tests for all the functions, and changed their
parameters. They now take a char* and a length to avoid using strlen
and other string functions.